맵 구현, 스테이지 소스파일 추가, 스테이지 헤더파일 수정
