+++ stage1.txt
@@ -1,0 +1,41 @@
+11111111111111111111111111111111111111111111111111
+10000000100000000000000010000000000100000000001001
+10700000102000000000004010000000000100000000001001
+10000000100000000000000010000000000100000000001001
+10000000100011111000111111110001111100011110001001
+10000600000010001000100000000000000100000010001001
+10000000000010501000100000000000000100000010001001
+10000000000010001000100000000000000100000010001001
+11111111111110001000100011111111000111100010001001
+10000000000000000000100010000000000100000010001001
+10000000000000000000100010200000000100000010001001
+10000000000000000000100010000000000100000010001001
+10001111100011111111100011111111111100011110001001
+10001000100010000000000000001000000000010000001001
+10001000100010000000000000001000000000010200001001
+10001000100010000000000000001000000000010000001001
+10001000100010001111111110001000111111111111111111
+10001000000000000000100010001000100000000000000001
+10901000000000000000104010001000100000000000000001
+10001000000000000000100010001000100000000000000001
+11111111111111111000100010001000100011111111110001
+10000000000000000000100010001000100010001000010001
+10000000000000000000100010001000100010801000010001
+10000000000000000000100000000000100000001000010001
+10001111100011111111100000000000100000001000010001
+10001000100010001000100000000000100000001000010001
+10001040100010001000100011111111111111111000010001
+10001000100010001000100010000000000000000000010001
+10001000100010001000100010000000000000000000010001
+10000000100010001000000010000000000000000000010001
+10000000100010001000000010001111100010001000010001
+10000000100010001000000010001000100010001000010001
+10001111100010001111100010001000100010001000010001
+10001000000000001000100010000000100010001000010001
+10001000000000001020100010000040100010001000010001
+10001000000000001000100010000000100010001000010001
+10001000111111111000100011111111100010001111110001
+10001000000000000000100000000000000010000000000001
+10001000000000000000100000000000000010000000000001
+10001000000000000000100000000000000010000000000001
+11111111111111111111111111111111111111111111111111
