+++ stage2.txt
@@ -1,0 +1,41 @@
+11111111111111111111111111111111111111111111111111
+10000000000100000000001000000000001000000000010001
+10700000000100000000001040000000001000000000010801
+10000000000100000000001000000000001000000000010001
+10001111000100011110001111100010001000111111110001
+10001000000100000010000000000010001000100000000001
+10001000000100000010000000000010001000100000000001
+10001000000100000010000000000010001000100000000001
+10001000111111100010001111111110001000100011111111
+10001000000100000010000000000010001000100000000001
+10001000000100000010000000000010501000100000000001
+10001000000100000010000000000010001000100000000001
+11111111000100011111111111100011111000111100010001
+10000000000100010000000000000000001000000000010001
+10000000000100010000000000000000201000000000010001
+10000000000100010000000000000000001000000000010001
+10001111111100010001111111100010001111111100010001
+10000000000000010000000000100010001000000100010001
+10200000000000010000000000100010001050000100010001
+10000000000000010000000000100010001000000100010001
+10001111111111111111111000111110001111000111110001
+10000000100000000000001000000010001000000100000001
+10000040100000000000001000000010001040000100000001
+10000000100000000000001000000010001000000100000001
+10001111100011110001111111100010001000111100011111
+10001000000000010000000000000010000000000100000001
+10001000000000010000000000000010000000000100000201
+10001000000000010000000000000010000000000100000001
+10001000100011110001111111111111111111111100010001
+10001000100000010001000000000000001000000000010001
+10501000100000010201000000000000001000000000010001
+10001000100000010001000000000000001000000000010001
+11111000111100011111000111111110001000111111110001
+10000000000100000000000100000000001000100000010001
+10000000020100000000000105000000001000100005010001
+10000000000100000000000100000000001000100000010001
+10001111111111111111000111111110001000100011110001
+10000000000100000000000000010000000000100000000001
+10000000090100000000000000010000000000100000000001
+10000000000100000000000000010000000000100000000001
+11111111111111111111111111111111111111111111111111
